input/vcd: Avoid truncation of shift to 32 bits.
(parse_contents): Cast constant 1 to uint64_t before shifting left.
Also fix whitespace errors.

+++ input/vcd.c
@@ -103,21 +103,21 @@ static gboolean read_until(FILE *file, GString *dest, char mode)
 				sr_err("Unexpected EOF, read started at %ld.", startpos);
 			return FALSE;
 		}
-		
+
 		if (mode == 'W' && g_ascii_isspace(c))
 			return TRUE;
-		
+
 		if (mode == 'N' && !g_ascii_isspace(c)) {
 			ungetc(c, file);
 			return TRUE;
 		}
-		
+
 		if (mode == '$') {
 			prev[0] = prev[1]; prev[1] = prev[2]; prev[2] = prev[3]; prev[3] = c;
 			if (prev[0] == '$' && prev[1] == 'e' && prev[2] == 'n' && prev[3] == 'd') {
 				if (dest != NULL)
 					g_string_truncate(dest, dest->len - 3);
-					
+
 				return TRUE;
 			}
 		}
@@ -135,29 +135,29 @@ static gboolean parse_section(FILE *file, gchar **name, gchar **contents)
 {
 	gboolean status;
 	GString *sname, *scontents;
-	
+
 	/* Skip any initial white-space */
 	if (!read_until(file, NULL, 'N')) return FALSE;
-	
+
 	/* Section tag should start with $. */
 	if (fgetc(file) != '$') {
 		sr_err("Expected $ at beginning of section.");
 		return FALSE;
 	}
-	
-	/* Read the section tag */	
+
+	/* Read the section tag */
 	sname = g_string_sized_new(32);
 	status = read_until(file, sname, 'W');
-	
+
 	/* Skip whitespace before content */
 	status = status && read_until(file, NULL, 'N');
-	
+
 	/* Read the content */
 	scontents = g_string_sized_new(128);
 	status = status && read_until(file, scontents, '$');
 	g_strchomp(scontents->str);
 
-	/* Release strings if status is FALSE, return them if status is TRUE */	
+	/* Release strings if status is FALSE, return them if status is TRUE */
 	*name = g_string_free(sname, !status);
 	*contents = g_string_free(scontents, !status);
 	return status;
@@ -187,7 +187,7 @@ static void remove_empty_parts(gchar **parts)
 			*dest++ = *src;
 		src++;
 	}
-	
+
 	*dest = NULL;
 }
 
@@ -204,7 +204,7 @@ static gboolean parse_header(FILE *file, struct context *ctx)
 
 	while (parse_section(file, &name, &contents)) {
 		sr_dbg("Section '%s', contents '%s'.", name, contents);
-	
+
 		if (g_strcmp0(name, "enddefinitions") == 0) {
 			status = TRUE;
 			break;
@@ -220,7 +220,7 @@ static gboolean parse_header(FILE *file, struct context *ctx)
 					sr_warn("Inexact rounding of samplerate, %" PRIu64 " / %" PRIu64 " to %" PRIu64 " Hz.",
 						q, p, ctx->samplerate);
 				}
-				
+
 				sr_dbg("Samplerate: %" PRIu64, ctx->samplerate);
 			} else {
 				sr_err("Parsing timescale failed.");
@@ -229,7 +229,7 @@ static gboolean parse_header(FILE *file, struct context *ctx)
 			/* Format: $var type size identifier reference $end */
 			gchar **parts = g_strsplit_set(contents, " \r\n\t", 0);
 			remove_empty_parts(parts);
-			
+
 			if (g_strv_length(parts) != 4)
 				sr_warn("$var section should have 4 items");
 			else if (g_strcmp0(parts[0], "reg") != 0 && g_strcmp0(parts[0], "wire") != 0)
@@ -246,17 +246,17 @@ static gboolean parse_header(FILE *file, struct context *ctx)
 				ctx->probes = g_slist_append(ctx->probes, probe);
 				ctx->probecount++;
 			}
-			
+
 			g_strfreev(parts);
 		}
-		
+
 		g_free(name); name = NULL;
 		g_free(contents); contents = NULL;
 	}
-	
+
 	g_free(name);
 	g_free(contents);
-	
+
 	return status;
 }
 
@@ -265,7 +265,7 @@ static int format_match(const char *filename)
 	FILE *file;
 	gchar *name = NULL, *contents = NULL;
 	gboolean status;
-	
+
 	file = fopen(filename, "r");
 	if (file == NULL)
 		return FALSE;
@@ -276,11 +276,11 @@ static int format_match(const char *filename)
 	 */
 	status = parse_section(file, &name, &contents);
 	status = status && (*name != '\0');
-	
+
 	g_free(name);
 	g_free(contents);
 	fclose(file);
-	
+
 	return status;
 }
 
@@ -316,23 +316,23 @@ static int init(struct sr_input *in, const char *filename)
 				return SR_ERR;
 			}
 		}
-		
+
 		param = g_hash_table_lookup(in->param, "downsample");
 		if (param) {
 			ctx->downsample = strtoul(param, NULL, 10);
 			if (ctx->downsample < 1)
 				ctx->downsample = 1;
 		}
-		
+
 		param = g_hash_table_lookup(in->param, "compress");
 		if (param)
 			ctx->compress = strtoul(param, NULL, 10);
-		
+
 		param = g_hash_table_lookup(in->param, "skip");
 		if (param)
 			ctx->skip = strtoul(param, NULL, 10) / ctx->downsample;
 	}
-	
+
 	/* Maximum number of probes to parse from the VCD */
 	ctx->maxprobes = num_probes;
 
@@ -342,12 +342,12 @@ static int init(struct sr_input *in, const char *filename)
 
 	for (i = 0; i < num_probes; i++) {
 		snprintf(name, SR_MAX_PROBENAME_LEN, "%d", i);
-		
+
 		if (!(probe = sr_probe_new(i, SR_PROBE_LOGIC, TRUE, name))) {
 			release_context(ctx);
 			return SR_ERR;
 		}
-			
+
 		in->sdi->probes = g_slist_append(in->sdi->probes, probe);
 	}
 
@@ -362,24 +362,24 @@ static void send_samples(const struct sr_dev_inst *sdi, uint64_t sample, uint64_
 	uint64_t buffer[CHUNKSIZE];
 	uint64_t i;
 	unsigned chunksize = CHUNKSIZE;
-		
+
 	if (count < chunksize)
 		chunksize = count;
 
 	for (i = 0; i < chunksize; i++)
 		buffer[i] = sample;
-	
+
 	packet.type = SR_DF_LOGIC;
-	packet.payload = &logic;	
+	packet.payload = &logic;
 	logic.unitsize = sizeof(uint64_t);
 	logic.data = buffer;
-	
+
 	while (count) {
 		if (count < chunksize)
 			chunksize = count;
-	
+
 		logic.length = sizeof(uint64_t) * chunksize;
-	
+
 		sr_session_send(sdi, &packet);
 		count -= chunksize;
 	}
@@ -389,20 +389,20 @@ static void send_samples(const struct sr_dev_inst *sdi, uint64_t sample, uint64_
 static void parse_contents(FILE *file, const struct sr_dev_inst *sdi, struct context *ctx)
 {
 	GString *token = g_string_sized_new(32);
-	
+
 	uint64_t prev_timestamp = 0;
 	uint64_t prev_values = 0;
-	
+
 	/* Read one space-delimited token at a time. */
 	while (read_until(file, NULL, 'N') && read_until(file, token, 'W')) {
 		if (token->str[0] == '#' && g_ascii_isdigit(token->str[1])) {
 			/* Numeric value beginning with # is a new timestamp value */
 			uint64_t timestamp;
 			timestamp = strtoull(token->str + 1, NULL, 10);
-			
+
 			if (ctx->downsample > 1)
 				timestamp /= ctx->downsample;
-			
+
 			/*
 			 * Skip < 0 => skip until first timestamp.
 			 * Skip = 0 => don't skip
@@ -423,9 +423,9 @@ static void parse_contents(FILE *file, const struct sr_dev_inst *sdi, struct con
 					/* Compress long idle periods */
 					prev_timestamp = timestamp - ctx->compress;
 				}
-			
+
 				sr_dbg("New timestamp: %" PRIu64, timestamp);
-			
+
 				/* Generate samples from prev_timestamp up to timestamp - 1. */
 				send_samples(sdi, prev_values, timestamp - prev_timestamp);
 				prev_timestamp = timestamp;
@@ -454,7 +454,7 @@ static void parse_contents(FILE *file, const struct sr_dev_inst *sdi, struct con
 			struct probe *probe;
 
 			bit = (token->str[0] == '1');
-		
+
 			g_string_erase(token, 0, 1);
 			if (token->len == 0) {
 				/* There was a space between value and identifier.
@@ -463,7 +463,7 @@ static void parse_contents(FILE *file, const struct sr_dev_inst *sdi, struct con
 				read_until(file, NULL, 'N');
 				read_until(file, token, 'W');
 			}
-			
+
 			for (i = 0, l = ctx->probes; i < ctx->probecount && l; i++, l = l->next) {
 				probe = l->data;
 
@@ -472,23 +472,23 @@ static void parse_contents(FILE *file, const struct sr_dev_inst *sdi, struct con
 				
 					/* Found our probe */
 					if (bit)
-						prev_values |= (1 << i);
+						prev_values |= (uint64_t)1 << i;
 					else
-						prev_values &= ~(1 << i);
-					
+						prev_values &= ~((uint64_t)1 << i);
+
 					break;
 				}
 			}
-			
+
 			if (i == ctx->probecount)
 				sr_dbg("Did not find probe for identifier '%s'.", token->str);
 		} else {
 			sr_warn("Skipping unknown token '%s'.", token->str);
 		}
-		
+
 		g_string_truncate(token, 0);
 	}
-	
+
 	g_string_free(token, TRUE);
 }
 
@@ -526,7 +526,7 @@ static int loadfile(struct sr_input *in, const char *filename)
 
 	/* Parse the contents of the VCD file */
 	parse_contents(file, in->sdi, ctx);
-	
+
 	/* Send end packet to the session bus. */
 	packet.type = SR_DF_END;
 	sr_session_send(in->sdi, &packet);
